refac: main.js -> main.ts 변경
- 테스트코드의 main 경로 수정: 렌터링 시간차에 따른 테스트 에러 이슈 해결
- app.ts 파일 삭제(ts 적용을 위한 경유지 역할)
