+++ index.html
@@ -1,5 +1,6 @@
 <!doctype html>
 <html lang="ko">
+
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
@@ -19,8 +20,10 @@
     };
   </script>
 </head>
+
 <body class="bg-gray-50">
-<div id="root"></div>
-<script type="module" src="/src/main.js"></script>
+  <div id="root"></div>
+  <script type="module" src="/src/main.ts"></script>
 </body>
+
 </html>
