+++ src/__tests__/cart.test.js
@@ -18,7 +18,7 @@ const addProductToCart = async (productName) => {
 
 beforeAll(async () => {
   document.body.innerHTML = '<div id="root"></div>';
-  await import("../main.js");
+  await import("../main.ts");
 });
 
 beforeEach(() => goTo("/"));
