--- src/main.js
@@ -1,17 +1,0 @@
-const enableMocking = () =>
-  import("./mocks/browser.js").then(({ worker }) =>
-    worker.start({
-      onUnhandledRequest: "bypass",
-    }),
-  );
-
-function main() {
-  import("./app.ts").catch((err) => console.error("Failed to load app.ts", err));
-}
-
-// 애플리케이션 시작
-if (import.meta.env.MODE !== "test") {
-  enableMocking().then(main);
-} else {
-  main();
-}
